requirements.txt
pyyaml==3.11
argcomplete==1.3.0
pep8==1.7.0